pyproject.toml
[project]
name = "phantommail"
version = "0.1.15"
description = "phantommail is an app that generates fake order data and customer emails. "
readme = "README.md"
authors = [
    { name = "Ben Selleslagh", email = "ben@dataframe.be" } 
]
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "faker>=35.2.0",
    "google-genai>=1.27.0",
    "langchain-google-genai>=2.0.9",
    "langchain>=0.3.18",
    "langgraph-cli[inmem]>=0.1.71",
    "langgraph>=0.2.70",
    "pandas>=2.3.1",
    "playwright>=1.54.0",
    "questionary>=2.1.1",
    "resend>=2.6.0",
]

[project.scripts]
phantommail = "phantommail:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "nest-asyncio>=1.6.0",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.5",
]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
    "D100",
    "D203",
    "D213",
]
lint.per-file-ignores."src/phantommail/cli/*" = ["T201"]
lint.per-file-ignores."src/phantommail/main.py" = ["T201", "D401"]

exclude = [
    "examples/*",
    "tests/*",
]
